Refine balancer config and split avatar settings
