+++ scripts/avatar.js
@@ -1,4 +1,4 @@
-import { AVATAR_PLACEHOLDER } from './config.js?v=2025-09-19-avatars-2';
+import { AVATAR_PLACEHOLDER } from './avatarConfig.js?v=2025-09-19-avatars-2';
 import { avatarNickKey } from './api.js?v=2025-09-19-avatars-2';
 import { renderAllAvatars } from './avatars.client.js';
