+++ scripts/avatarAdmin.js
@@ -1,7 +1,7 @@
 // scripts/avatarAdmin.js
 import { log } from './logger.js?v=2025-09-19-avatars-2';
 import { uploadAvatar, loadPlayers, avatarNickKey } from './api.js?v=2025-09-19-avatars-2';
-import { AVATAR_PLACEHOLDER, AVATAR_WORKER_BASE } from './config.js?v=2025-09-19-avatars-2';
+import { AVATAR_PLACEHOLDER, AVATAR_WORKER_BASE } from './avatarConfig.js?v=2025-09-19-avatars-2';
 import { reloadAvatars } from './avatars.client.js';
 
 const MAX_FILE_SIZE = 2 * 1024 * 1024;
