+++ scripts/avatarConfig.js
@@ -1,0 +1,4 @@
+export const AVATAR_WORKER_BASE = 'https://avatarsproxy.vartaclub.workers.dev';
+export const AVATAR_CACHE_BUST = '2025-09-19-avatars-2';
+export const AVATAR_PLACEHOLDER = 'assets/default_avatars/av0.png';
+export const ASSETS_VER = AVATAR_CACHE_BUST;
