+++ scripts/gameday.js
@@ -1,5 +1,5 @@
 import { log } from './logger.js?v=2025-09-19-avatars-2';
-import { AVATAR_PLACEHOLDER } from './config.js?v=2025-09-19-avatars-2';
+import { AVATAR_PLACEHOLDER } from './avatarConfig.js?v=2025-09-19-avatars-2';
 import { getPdfLinks, fetchOnce, CSV_URLS, avatarNickKey } from "./api.js?v=2025-09-19-avatars-2";
 import { rankLetterForPoints } from './rankUtils.js?v=2025-09-19-avatars-2';
 import { renderAllAvatars, reloadAvatars } from './avatars.client.js';
