+++ scripts/profile.js
@@ -3,7 +3,7 @@ import { getProfile, uploadAvatar, getPdfLinks, fetchPlayerGames, safeSet, safeG
 import { rankLetterForPoints } from './rankUtils.js?v=2025-09-19-avatars-2';
 import * as Avatars from './avatars.client.js';
 import { noteAvatarFailure, setImgSafe, updateInlineAvatarImages } from './avatarAdmin.js?v=2025-09-19-avatars-2';
-import { AVATAR_PLACEHOLDER } from './config.js?v=2025-09-19-avatars-2';
+import { AVATAR_PLACEHOLDER } from './avatarConfig.js?v=2025-09-19-avatars-2';
 
 let gameLimit = 0;
 let gamesLeftEl = null;
